requirements.txt
streamlit
yt-dlp
openai-whisper
torch
ffmpeg